♻️ refactor(scheduler,channel): port to PG sqlc types
Refs: #482

+++ internal/scheduler/service.go
@@ -640,7 +640,7 @@ func (s *Service) ListJobRuns(ctx context.Context, jobID string, limit int) ([]J
 	rows, err := s.q.ListSchedJobRuns(ctx, sqlc.ListSchedJobRunsParams{
 		JobID:  jobID,
 		UserID: nil,
-		Limit:  int64(limit),
+		Limit:  int32(limit),
 		Offset: 0,
 	})
 	if err != nil {
